streamlit/chatbot/pyproject.toml
[project]
name = "streamlit-chatbot"
version = "0.1.0"
description = "A Streamlit chatbot supporting multiple LLM providers"
requires-python = ">=3.12,<3.13"
dependencies = [
    "streamlit>=1.45.0",
    "openai>=1.78.0",
    "anthropic>=0.51.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.27",
    "pydantic>=2.6.1",
    "tomli>=2.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "db", "models"]

[tool.ruff]
line-length = 88
target-version = "py312" 
